clinic/requirements.txt
asgiref==3.11.0
Django==5.2.8
djangorestframework==3.16.1
